watson/buddybot/requirements.txt
pyyaml
requests
SpeechRecognition
pytest
PyAudio
